Se cambia modelo

+++ WebApplication1/Models/Games.cs
@@ -19,13 +19,13 @@ namespace WebApplication1.Models
         public string State { get; set; }
 
         [Required]
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         [Required]
-        public double Percent_Rent { get; set; }
+        public decimal Percent_Rent { get; set; }
 
         [Required]
-        public double Reward_Cooler_Coins { get; set; }
+        public decimal Reward_Cooler_Coins { get; set; }
 
         [Required]
         public string Image { get; set; }
